requirements.txt
rsa==4.9
pycryptodome==3.19.0